Se aplicó estilo al botón de cerrar sesión y se insertó dentro del sidebar
